Refactor gitignore to blacklist
Originally, .gitignore was a whitelist: it began by excluding
everything, then selectively including those sub-directories that
were to be included. There are at least two issues with this
appraoch.

First, a bug in the implementation would lead to files not being
seen as changed potentially resulting in change loss. This is more
than theoretical. The original implementation of subdir/** does
not work as advertised: creating a file in test/fixtures/ (for
example) would not see that file as new. The reason is from the
git docs:

> It is not possible to re-include a file if a parent directory of
> that file is excluded.
-- https://stackoverflow.com/a/35569180/2908724

Second, this pattern is rather uncommon practice in my
experience and may lead to confuse when editing.