Fix Python dependency errors for mkdocs in Docker
There exists some as-yet-undocumented edge-case that prevents Debian 9,
which node image is based off of, from cleanly installing mkdocs with
Python3.

Several versions of pip3 were attempted, as well as a few versions of
Python3.x. All seemed to raise a dependencyError with livereload and
tornado.

This approach is perhaps a bit hacky, but is honest about the upstream
dependency. If for some reason `mkdocs` should fail to work when using
this image, users are directed to use the squidfunk/mkdocs-material
docker image directly.
Dockerfile fixes