Configure dependabot to perform version updates
This will cause dependabot to open PRs on a daily cadence when
dependencies are updated in NPM. We want this especially so that we
don't miss picking up published changes to
@jupiterone/security-policy-templates.

Additionally, check for Dockerfile base image updates once a week.
These may contain important security updates, so let's stay on top of
that, too.