Refactor workflow; publish to Docker Hub
This flow should publish to NPM any time there is a tagged version
commit at HEAD of branch.

This flow should publish to Docker Hub any time the Dockerfile changes,
and/or any time there has been an NPM publish.

NOTE: it is likely that the conditionals in this flow could be avoided
by making use of GitHub Releases, and relying on the `release` workflow
event. That might be worth exploring down the road a bit. See:
https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github